Files.find() -> This method is similar to Files.walk(), but takes an additional argument of type BiPredicate that is used to filter the files and directories.
Remember that a BiPredicate takes two arguments and returns a boolean. In this case:

The first argument is the Path object that represents the file or directory.
The second argument is a BasicFileAttributes object that represents the attributes of the file or directory in the file system (like creation time, if it's a file, directory or symbolic link, size, etc.).
The returned boolean indicates if the file should be included in the returned stream.